Remove default nav sections and user
